Add files via upload
https://github.com/msanders13/Class2Assignment2/
I really had trouble dealing with the unordered list I created for my hobbies, I could not get the list to center without moving the bullets. I was able to underline, put a background on and move "My hobbies include" and able to center the ul but not everything together at once.